Jim Updates 5.8.20
Hey guys, I made some changes in the data files to combine them in my pre-work file. I used bind_rows to combine the 06, 08, and 18 surveys. There were a couple of other data sets in the folder I wasn't sure what to do with, but once we finalize the data sets we're using, I think we should use map_df to combine everything. I also played around making some descriptive tables with gt. This code is very repetitive, so I need to find a way to work it into a function. I then did some practicing using prrr just to try to get more comfortable with it. I'm going to try to keep doing that over the next couple of days.